requirements.txt
ipython==8.4.0
kubernetes==23.6.0
pytest==7.1.2
python-dotenv==0.20.0
pytz==2021.1
